Remove use of HAL and MIDI
MIDI over USB would be useful for debugging but the complexity of its
crates is not worth it. It also depends on community HAL that is
unfortunately missing many features that will be needed here. Because of
this, I'm dropping all these dependencies. HALs will be written only per
the code used and needed in this project.

Signed-off-by: Petr Horáček <phoracek@redhat.com>