Don't run disabled degrees
The original caused where an oscillator was disabled by setting its
frequency to 0 had a side effect - the oscillator kept adding DC offset,
based on the phase it got stuck in. Due to this, the center of the
output was moving each time the instrument was switched from chord to
single-voice mode. This patch fixes this by adding an explicit
enabled/disabled parameter to instrument degrees.

Signed-off-by: Petr Horacek <hrck@protonmail.com>