Move all board resources behind an abstraction
With this patch, all the board resources, both internal (flash) and
external (pots, cvs, audio, leds), are moved behind a "system"
abstraction. This abstraction does not care about usage of these
components, it merely exposes them for user to leverage. Interface that
is returned is a structure with public fields, allowing for split of the
structure into multiple subroutines.

Signed-off-by: Petr Horacek <hrck@protonmail.com>