Change V/OCT inputs to scale 0.0 to 10.0 V
With this, 10Vpp is used for both modulation and V/OCT inputs. It also
makes it clear unidirectional, adjusts with beatstep and allows us to go
all the way to the lowest C since it starts on whole number.

There is no clear standard, but -5 to 5V and 0 to 10V seems more common
than other ranges and I also like the numbers!

Signed-off-by: Petr Horacek <hrck@protonmail.com>