Introduce CV calibration mode
When the button it held down and a jack gets plugged to the V/OCT input,
the module enters calibration mode. In that, C should be played on the
input device, then button clicked, then C one octave higher and one
click again. After that, calibration is done and input keys should be
preciselly mapped and quantized.

Signed-off-by: Petr Horacek <hrck@protonmail.com>