Introduce V/OCT control of chords
When chord CV is connected while holding the button, the module enters
calibration. After that is successfully done, chords are controled
through V/OCT by white keys between 1 and 4 V (when chord pot is on
zero) or 2 and 5 V (when chord pot is on full). This mode is exited when
chord CV is disconnected.

Signed-off-by: Petr Horacek <hrck@protonmail.com>