Remove amplitude balancing on small chords
With this, every oscillator is given the same amplitude. This guarantees
that all tones will keep their continuous loudness and with that fixes
an issue where change of size of a chord would weaken continuous tones
kept between chords, e.g. the root. The downside of this is that small
chords and solo tones have much weaker signal.

Signed-off-by: Petr Horacek <hrck@protonmail.com>