Adjust V/oct response to reflect new voltage range
Now with the V/oct range being -5 to +5 V, it becomes little more
difficult to track with the Arturia Beat Step Pro, which is better
suited for 0 to 10 V.

This patch extends the offset range for NOTE and lowers the octaves of
NOTE'. However, the best results would be obtained by offseting BSP
with a mixer.

Signed-off-by: Petr Horacek <hrck@protonmail.com>