Add firmware without white key quantization
Due to white key quantization, the module is quite sensitive to
inaccurate CV input of TONE. This is especially notable when the same CV
is simultaneously used to control the pitch of another module as well.

This patch adds a feature flag `even_quantization`. If this is enabled,
white key quantization is switched off.

Signed-off-by: Petr Horacek <petr@zlosynth.com>