Disable white key quantization by default
It seems that most people use the module with a regular CV input, not
with keyboards. While this quantization feature can be fun, it can make
control using a regular CV, since some of the notes only subscribe to
a narrow part of the 1V/oct range.

With this patch, the default is now to snap to the closest note, leaving
the white key quantization as an alternative firmware.

Signed-off-by: Petr Horacek <petr@zlosynth.com>